requirements.txt
# Core ML & Vision
ultralytics
opencv-python
numpy
Pillow
torch
pyyaml

# API & Environment
python-dotenv
requests

# Utilities
tqdm
pathlib
colorama

# Error handling & retries
backoff